chore : refactoring typography (making it pure)
